Solucionar conflictos
